% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_market_price.R
\name{aggregate_market_price}
\alias{aggregate_market_price}
\title{Aggregate market price}
\usage{
aggregate_market_price(
  market_prices,
  unit = "15 mins",
  aggregate_price_fun = mean()
)
}
\arguments{
\item{market_prices}{Data frame containing the market prices.}

\item{unit}{Character string specifying a time unit or a multiple of a
unit to be rounded to. Valid base units are second, minute, hour, day,
week, etc. See \code{lubridate::\link[lubridate:round_date]{round_date}}.}

\item{aggregate_price_fun}{Function to use to aggregate prices within
the round unit. Default to \code{mean}.}
}
\value{
The data frame of aggregated market prices.
}
\description{
Aggregate the data frame of market prices with a
specific aggregate function within a time interval.
}
\references{
H. Shefrin & M. Statman, 1985
}
\seealso{
\code{\link{closest_market_price}},
\code{lubridate::\link[lubridate:round_date]{round_date}}
}
\author{
L. Mazzucchelli & M. Zanotti
}
