% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{checks}
\alias{checks}
\alias{check_gainloss}
\title{Checks functions}
\usage{
check_gainloss(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
A character string containing an error message or NULL.
}
\description{
Functions to perform checks on arguments.
}
\section{Functions}{
\itemize{
\item \code{check_gainloss}: Check consistency of gains and losses "values" results.
}}

\keyword{internal}
