% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{client_4273N}
\alias{client_4273N}
\title{Client 4273N financial transactions}
\format{
A data frame with 19 rows and 6 variables:
\describe{
\item{client}{id of the investor}
\item{type}{binary variable indicating the type of operation, B = buy and S = sell}
\item{asset}{id of the traded asset}
\item{qty}{quantity of the traded asset}
\item{prz}{market price of the traded asset}
\item{datetime}{timestamp of the operation}
}
}
\source{
\url{https://www.directa.it/index-ita.html}
}
\usage{
client_4273N
}
\description{
A sample dataset containing 19 transactions over time.
}
\keyword{datasets}
