% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_market_price.R
\name{closest_market_price}
\alias{closest_market_price}
\title{Closest market price}
\usage{
closest_market_price(asset, datetime, market_prices)
}
\arguments{
\item{asset}{Character name of the asset to look for.}

\item{datetime}{POSIXct of the datetime at which looking for the asset's
price.}

\item{market_prices}{Data frame containing the market prices.}
}
\value{
The value of the asset's price.
}
\description{
Find the market price closest to a certain datetime and for
as specific asset.
}
\details{

}
\references{
H. Shefrin & M. Statman, 1985
}
\seealso{
\code{\link{evaluate}}
}
\author{
L. Mazzucchelli & M. Zanotti
}
