% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispositionEffect-package.R
\docType{package}
\name{dispositionEffect-package}
\alias{dispositionEffect-package}
\title{dispositionEffect: A package that allows to perform analyses on disposition effect}
\description{
The dispositionEffect package allows to perform analyses on the
disposition effect using financial market data.
}
\section{dispositionEffect functions}{

\describe{
\item{disposition_effect}{Compute the disposition effect based on
realized and paper gains and losses}
}
}

