% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-operators.R
\name{ewise_mean}
\alias{ewise_mean}
\title{Element-wise mean calculation}
\usage{
ewise_mean(x, y, na.rm = FALSE, zero.substitute = FALSE)
}
\arguments{
\item{x, y}{Numeric vectors of the same length.}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}

\item{zero.substitute}{Logical. If TRUE zero values are substituted with
corresponding non-missing values whether possible.}
}
\value{
Numeric vector of averages.
}
\description{
The function allows for the calculation of the mean on two
vectors element-wise.
}
\examples{
  \dontrun{
  x <- 1:10
  y <- 99:90
  ewise_mean(x, y)

  x[1] <- NA
  ewise_mean(x, y, na.rm = TRUE)

  x[1] <- 0
  ewise_mean(x, y, zero.substitute = TRUE)
  }

}
\keyword{internal}
