% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package.R
\docType{data}
\name{investor}
\alias{investor}
\title{Sample investor financial transactions}
\format{
A data frame with 19 rows and 6 variables:
\describe{
\item{investor}{id of the investor}
\item{type}{binary variable indicating the type of operation, B = buy and S = sell}
\item{asset}{id of the traded asset}
\item{quantity}{quantity of the traded asset}
\item{price}{market price of the traded asset}
\item{datetime}{timestamp of the operation}
}
}
\source{
\url{https://www.directa.it/index-ita.html}
}
\usage{
investor
}
\description{
A sample dataset containing 19 transactions over time.
}
\keyword{datasets}
