% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_initial_data.R
\name{manipulate_initial_data}
\alias{manipulate_initial_data}
\alias{aggregate_transactions}
\alias{reduce_transactions}
\alias{clean_tansactions}
\alias{aggregate_market_prices}
\alias{subset_market_prices}
\alias{clean_market_prices}
\title{Manipulate market prices}
\usage{
aggregate_transactions(portfolio_transactions)

reduce_transactions(portfolio_transactions, unit = "1 mins")

clean_tansactions(portfolio_transactions, unit = "1 mins")

aggregate_market_prices(
  market_prices,
  unit = "15 mins",
  aggregate_price_fun = mean
)

subset_market_prices(market_prices, portfolio_transactions, unit = "15 mins")

clean_market_prices(
  market_prices,
  portfolio_transactions,
  unit = "15 mins",
  aggregate_price_fun = mean
)
}
\arguments{
\item{portfolio_transactions}{Data frame. The investor's transactions data frame.}

\item{unit}{Character string specifying a time unit or a multiple of a
unit to be rounded to. Valid base units are second, minute, hour, day,
week, etc. See \code{lubridate::\link[lubridate:round_date]{round_date}}.}

\item{market_prices}{Data frame containing the market prices.}

\item{aggregate_price_fun}{Function to use to aggregate prices within
the round unit. Default to \code{mean}.}
}
\value{
The data frame of aggregated market prices.
}
\description{
Aggregate the data frame of market prices with a
specific aggregate function within a time interval.
}
\section{Functions}{
\itemize{
\item \code{aggregate_transactions}: Aggregate the data frame of investors'
transactions that happened at the same datetime.

\item \code{reduce_transactions}: Reduce the data frame of investors'
transactions by aggregating those transactions that happened within
a time interval.

\item \code{clean_tansactions}: Clean the data frame of investors'
transactions.

\item \code{aggregate_market_prices}: Aggregate the data frame of market prices
with a specific aggregate function within a time interval.

\item \code{subset_market_prices}: Subset the data frame of market prices
based on investor's traded assets and datetimes of transactions.

\item \code{clean_market_prices}: Clean the data frame of market prices.
}}

\references{
H. Shefrin & M. Statman, 1985
}
\seealso{
\code{\link{closest_market_price}},
\code{lubridate::\link[lubridate:round_date]{round_date}}
}
\author{
L. Mazzucchelli & M. Zanotti
}
