% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package.R
\docType{data}
\name{marketprices}
\alias{marketprices}
\title{Market prices of assets traded by the sample investor}
\format{
A data frame with 6895 rows and 4 variables:
\describe{
\item{asset}{id of the asset}
\item{datetime}{timestamp of market price}
\item{price}{market price of the asset}
}
}
\source{
\url{https://www.directa.it/index-ita.html}
}
\usage{
marketprices
}
\description{
A sample dataset containing 6895 market prices of 5 different assets over time.
}
\keyword{datasets}
