% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{marketprices_4273N}
\alias{marketprices_4273N}
\title{Market prices of assets traded by client 4273N}
\format{
A data frame with 6895 rows and 4 variables:
\describe{
\item{asset}{id of the asset}
\item{datetime}{timestamp of market price}
\item{prz}{market price of the asset}
}
}
\source{
\url{https://www.directa.it/index-ita.html}
}
\usage{
marketprices_4273N
}
\description{
A sample dataset containing 6895 market prices of 5 different assets over time.
}
\keyword{datasets}
