% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper_compute}
\alias{paper_compute}
\alias{paper_count}
\alias{paper_total}
\alias{paper_value}
\alias{paper_duration}
\title{Papers' estimation}
\usage{
paper_count(portf_qty, portf_prz, hist_prz, allow_short = FALSE)

paper_total(portf_qty, portf_prz, hist_prz, allow_short = FALSE)

paper_value(portf_qty, portf_prz, hist_prz, allow_short = FALSE)

paper_duration(
  portf_qty,
  portf_prz,
  hist_prz,
  last_dtt,
  trx_dtt,
  allow_short = FALSE
)

paper_compute(
  portf_qty,
  portf_prz,
  hist_prz,
  last_dtt,
  trx_dtt,
  assets,
  allow_short = FALSE,
  method = "all"
)
}
\arguments{
\item{portf_qty}{Numeric vector. The portfolio quantities of assets into the
investor's portfolio.}

\item{portf_prz}{Numeric vector. The portfolio prices of assets into the
investor's portfolio.}

\item{hist_prz}{Numeric vector. The market prices of assets into the
investor's portfolio.}

\item{allow_short}{Logical. If TRUE short positions are allowed, otherwise only
long positions are allowed.}

\item{last_dtt}{POSIXct vector. The portfolio date-times of assets into the
investor's portfolio.}

\item{trx_dtt}{POSIXct vector. The market date-times at which the transaction
is going to occur.}

\item{assets}{Character vector. The name of assets into the investor's
portfolio.}

\item{method}{Character string. The method used to compute papers.
Allowed values are "count", "total", "value", "duration" and "all".}
}
\value{
The described functions have different return behaviours.

\describe{
\item{\code{paper_compute}}{returns a \link[tibble:tibble-package]{tibble}
containing the values of paper gains and paper losses computed by
means of the chosen method on each portfolio assets.}
\item{\code{paper_count}}{returns a named vector containing the values
of paper gains and paper losses computed using the count method.}
\item{\code{paper_total}}{returns a named vector containing the values
of paper gains and paper losses computed using the total method.}
\item{\code{paper_value}}{returns a named vector containing the values
of paper gains and paper losses computed using the value method.}
\item{\code{paper_duration}}{returns a named vector containing the values
of paper gains and paper losses computed using the duration method.}
}

In particular:

\itemize{
\item{\code{RG_"method"}}{contains Realized Gains results.}
\item{\code{RL_"method"}}{contains Realized Losses results.}
\item{\code{PG_"method"}}{contains Paper Gains results.}
\item{\code{PL_"method"}}{contains Paper Losses results.}
}
}
\description{

}
\details{

}
\section{Functions}{
\itemize{
\item \code{paper_count}: Computation of paper gains and paper losses as
simple counts.

\item \code{paper_total}: Computation of paper gains and paper losses as
simple quantity of assets.

\item \code{paper_value}: Computation of paper gains and paper losses as
simple expected return of assets.

\item \code{paper_duration}: Computation of paper gains and paper losses as
simple financial duration.

\item \code{paper_compute}: Wrapper that calls other paper_. functions to
compute paper gains and paper losses based on the chosen method.
}}

\references{
H. Shefrin & M. Statman, 1985
}
\seealso{
\code{\link{realized_compute}}, \code{\link{gains_and_losses}}
}
\author{
L. Mazzucchelli & M. Zanotti
}
