% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package.R
\docType{data}
\name{portfolio_results}
\alias{portfolio_results}
\title{Realized and paper results}
\format{
A data frame with 5 rows and 21 variables:
\describe{
\item{investor}{id of the investor}
\item{asset}{id of the traded asset}
\item{quantity}{quantity of the traded asset at the end of the portfolio
updating process}
\item{price}{last market price of the traded asset}
\item{datetime}{timestamp of the last operation}
\item{RG_count}{realized gains via count method}
\item{RL_count}{realized losses via count method}
\item{PG_count}{paper gains via count method}
\item{PL_count}{paper losses via count method}
\item{RG_total}{realized gains via total method}
\item{RL_total}{realized losses via total method}
\item{PG_total}{paper gains via total method}
\item{PL_total}{paper losses via total method}
\item{RG_value}{realized gains via value method}
\item{RL_value}{realized losses via value method}
\item{PG_value}{paper gains via value method}
\item{PL_value}{paper losses via value method}
\item{RG_duration}{realized gains via duration method}
\item{RL_duration}{realized losses via duration method}
\item{PG_duration}{paper gains via duration method}
\item{PL_duration}{paper losses via duration method}
}
}
\source{
\url{https://www.directa.it/index-ita.html}
}
\usage{
portfolio_results
}
\description{
Results obtained by means of \code{portfolio_compute} on the
data sets \code{investor} and \code{marketprices}.
}
\keyword{datasets}
