% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_update.R
\name{portfolio_update}
\alias{portfolio_update}
\title{Portfolio Compute}
\usage{
portfolio_update(
  portfolio_transactions,
  market_prices,
  method = "all",
  allow_short = FALSE,
  time_threshold = "5 mins",
  posneg_portfolios = FALSE,
  portfolio_statistics = FALSE,
  verbose = c(0, 0),
  progress = TRUE
)
}
\arguments{
\item{portfolio_transactions}{Data frame. The investor's transactions data frame.}

\item{market_prices}{Data frame containing the market prices.}

\item{method}{Character string containing the method to use to compute
realized and paper gains and losses. If "none" nothing is computed but the
investor's portfolio update. Otherwise it has to be one of "count", "total",
"value", "duration" and "all".}

\item{allow_short}{Logical. If TRUE short positions are allowed, otherwise only
long positions are allowed.}

\item{time_threshold}{character in the format "value units" indicating the
time threshold at which the computed financial difftime has to be evaluated
(for instance "05 mins" or "20 hours").
The allowed units are "secs", "mins", "hours", "days" and "weeks"
(See \code{base::\link[base:difftime]{difftime}}).}

\item{posneg_portfolios}{Logical. If TRUE the realized and paper gains and
losses for the positive (that is when the investor's portfolio value, as
computed through \code{\link{evaluate_portfolio}}, is greater than zero)
and the negative (that is when the investor's portfolio value, as computed
through \code{\link{evaluate_portfolio}}, is smaller than zero) portfolios
are returned.}

\item{portfolio_statistics}{Logical. If TRUE some statistical indexes are computed
on the portfolio and returned.}

\item{verbose}{Numeric or logical vector of length 2 that allows to control
for the function verbosity.}

\item{progress}{Logical. If TRUE a progress bar is displayed.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the investor's
portfolio and the values of realized and paper gains and losses
computed by means of the chosen method on each portfolio assets.
}
\description{
Computation of all the transaction updates and the
realized and paper gains and losses for each assets.
}
\details{

}
\references{
H. Shefrin & M. Statman, 1985
}
\seealso{
\code{\link{realized_compute}}, \code{\link{paper_compute}},
\code{\link{portfolio_update}}
}
\author{
L. Mazzucchelli & M. Zanotti
}
