% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{checks}
\alias{checks}
\alias{check_values}
\alias{check_gainloss}
\title{Checks functions}
\usage{
check_values(
  input_values,
  target_values,
  no_exception = FALSE,
  weak_target = FALSE
)

check_gainloss(df)
}
\arguments{
\item{input_values}{Input to check against target.}

\item{target_values}{Target to check for in input.}

\item{no_exception}{Logical. If TRUE different values of inputs from targets
are not allowed to exist.}

\item{weak_target}{Logical. If TRUE inputs do not need to contain all the
targets.}

\item{df}{Data frame}
}
\description{
Functions to perform checks on arguments.
}
\section{Functions}{
\itemize{
\item \code{check_values}: Check consistency of values against targets and vice versa.

\item \code{check_gainloss}: Check consistency of gains and losses "values" results.
}}

\keyword{internal}
