% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime.R
\name{difference_in_time}
\alias{difference_in_time}
\alias{financial_difftime}
\alias{compare_difftime}
\title{Difference in time}
\usage{
financial_difftime(from, to, pre_market = 8, after_market = 22)

compare_difftime(from, to, time_threshold = "0 mins")
}
\arguments{
\item{from}{POSIXct for the initial date.}

\item{to}{POSIXct for the final date.}

\item{pre_market}{numeric or character representing the hour of the day at
which the pre-market begins.}

\item{after_market}{numeric or character representing the hour of the day at
which the after-market ends.}

\item{time_threshold}{character in the format "value units" indicating the
time threshold at which the computed financial difftime has to be evaluated
(for instance "05 mins" or "20 hours").
The allowed units are "secs", "mins", "hours", "days" and "weeks"
(See \code{base::\link[base:difftime]{difftime}}).}
}
\value{
\describe{
\item{\code{financial_difftime}}{returns a numeric value in hours.
To convert in a different time unit use
\code{base::\link[base:as.numeric]{as.numeric}} and specify the
\code{units} argument (units = c("secs", "mins", "hours", "days",
"weeks")).}
\item{\code{compare_difftime}}{returns a character string whose possible
values are "greater" or "smaller", indicating whether the
\code{financial_difftime} evaluated between the two dates is greater or
smaller than the chosen threshold.}
}
}
\description{
Difference in time
}
\details{

}
\section{Functions}{
\itemize{
\item \code{financial_difftime}: Extension of the
\code{base::\link[base:difftime]{difftime}} function to calculate the
actual financial difftime taking into account pre and after market
moments and financial markets closing days.

\item \code{compare_difftime}: Comparison of the \code{financial_difftime}
with respect to a given time threshold.
}}

\examples{
  from <- as.POSIXct("2020-01-01 09:00:00", tz = "UTC")
  to <- as.POSIXct("2020-01-01 18:00:00", tz = "UTC")
  difftime(to, from)
  as.numeric(difftime(to, from))
  financial_difftime(from, to)

  from <- as.POSIXct("2020-01-01 09:00:00", tz = "UTC")
  to <- as.POSIXct("2020-01-02 18:00:00", tz = "UTC")
  difftime(to, from)
  as.numeric(difftime(to, from), units = "hours")
  financial_difftime(from, to)
  financial_difftime(from, to, after_market = 19) # Italian market

  compare_difftime(from, to, "5 mins")
  compare_difftime(from, to, "5 hours")
  compare_difftime(from, to, "5 days")

}
\references{
H. Shefrin & M. Statman, 1985
}
\seealso{
\code{base::\link[base:difftime]{difftime}}
}
\author{
L. Mazzucchelli & M. Zanotti
}
