% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disposition_effect.R
\name{disposition_effect}
\alias{disposition_effect}
\alias{disposition_difference}
\title{Disposition Effect}
\usage{
disposition_effect(realized_gains, paper_gains, realized_losses, paper_losses)

disposition_difference(gains, losses)
}
\arguments{
\item{realized_gains}{Numeric vector (or scalar) containing realized gains
values.}

\item{paper_gains}{Numeric vector (or scalar) containing paper gains
values.}

\item{realized_losses}{Numeric vector (or scalar) containing realized losses
values.}

\item{paper_losses}{Numeric vector (or scalar) containing paper losses
values.}

\item{gains}{Numeric vector (or scalar) containing gains.}

\item{losses}{Numeric vector (or scalar) containing losses.}
}
\value{
Numeric vector (or scalar) with the value(s) of disposition
effect(s) or disposition difference(s).
}
\description{
Compute the disposition effect in a vectorized way.
}
\details{
The disposition effect is defined as
\eqn{DE = (Realized Gain / (Realized Gain - Paper Gain)) -
       (Realized Loss / (Realized Loss + Paper Loss))}

The disposition difference is defined as
\eqn{DD = Realized Gain - |Realized Loss|}
or
\eqn{DD = Paper Gain - |Paper Loss|}
}
\section{Functions}{
\itemize{
\item \code{disposition_effect}: Compute the disposition effect as defined
by L. Mazzucchelli et al.

\item \code{disposition_difference}: Compute the disposition difference as defined
by L. Mazzucchelli et al.
}}

\examples{
  data(portfolio_results)

  # Disposition effect on count for each traded asset
  de <- disposition_effect(portfolio_results$RG_count,
                           portfolio_results$PG_count,
                           portfolio_results$RL_count,
                           portfolio_results$PL_count)
  names(de) <- portfolio_results$asset

  # Average disposition effect of the investor
  mean(de)

  # Realized difference on value for each asset
  disposition_difference(portfolio_results$RG_value, portfolio_results$RL_value)

  # Paper difference on value for each asset
  disposition_difference(portfolio_results$PG_value, portfolio_results$PL_value)

}
\references{
H. Shefrin & M. Statman, 1985
}
\author{
L. Mazzucchelli & M. Zanotti
}
