% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate_portfolio}
\title{Portfolio evaluation}
\usage{
evaluate_portfolio(portfolio_df, trx_date, df_asset_prices, statistics = FALSE)
}
\arguments{
\item{portfolio_df}{Data frame of the investor's portfolio at time t.}

\item{trx_date}{POSIXct value of the transaction date.}

\item{df_asset_prices}{Data frame containing the market prices.}

\item{statistics}{Logical. If TRUE some statistical indexes are computed
on the portfolio and returned.}
}
\value{
The portfolio value as the sum of each asset quantity times the
excess return of each asset with respect to the market.
}
\description{
Portfolio evaluation
}
\details{

}
\references{
H. Shefrin & M. Statman, 1985
}
\seealso{
\code{\link{portfolio_update}}, \code{\link{gains_and_losses}},
\code{\link{closest_historical_price}}
}
\author{
L. Mazzucchelli & M. Zanotti
}
