% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate_portfolio}
\title{Portfolio evaluation}
\usage{
evaluate_portfolio(portfolio, market_prices)
}
\arguments{
\item{portfolio}{Data frame of the investor's portfolio at time t.}

\item{market_prices}{Data frame containing the market prices.}
}
\value{
The portfolio value.
}
\description{
Calculate the portfolio value as the sum of each
asset portfolio quantity times the excess return of each asset
with respect to the market.
}
\seealso{
\code{\link{portfolio_compute}}, \code{\link{gains_losses}},
\code{\link{closest_market_price}}
}
