% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializers.R
\name{initializers}
\alias{initializers}
\alias{initializer_portfolio}
\alias{initializer_results}
\title{Initializers functions}
\usage{
initializer_portfolio(client, client_assets)

initializer_results(client, client_assets, method = "all")
}
\arguments{
\item{client}{Character string. The name of the investor.}

\item{client_assets}{Character vector. The names of the assets traded by
the investor.}

\item{method}{Character string. The method used to compute papers.
Allowed values are "count", "total", "value", "duration" and "all".}
}
\value{
Empty \link[tibble:tibble-package]{tibble} of either investor's
portfolio or investor's realized and paper gains and losses results
based on the chosen method.
}
\description{
Initializers functions
}
\section{Functions}{
\itemize{
\item \code{initializer_portfolio}: Empty \link[tibble:tibble-package]{tibble}
of investor's portfolio.

\item \code{initializer_results}: Empty \link[tibble:tibble-package]{tibble}
of investor's realized and paper gains and losses results based on
the chosen method.
}}

\author{
L. Mazzucchelli & M. Zanotti
}
\keyword{internal}
