% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-operators.R
\name{negative_match}
\alias{negative_match}
\alias{\%!in\%}
\title{Negative match operator}
\usage{
lhs \%!in\% rhs
}
\description{
See \code{base::\link[base:match]{\%in\%}} for details on the
match operator.
The negative match operator is simply the negation of the match operator
obtained via the \code{base::\link[base:Negate]{Negate}} function.
}
\keyword{internal}
