% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{updates}
\alias{updates}
\alias{prz_update}
\alias{dtt_update}
\alias{results_update}
\alias{meanvalue_compute}
\title{Update investor's portfolio or realized and paper results}
\usage{
prz_update(
  portfolio_quantity,
  portfolio_price,
  transaction_quantity,
  transaction_price,
  transaction_type
)

dtt_update(
  portfolio_quantity,
  portfolio_datetime,
  transaction_quantity,
  transaction_datetime,
  transaction_type
)

results_update(realized_and_paper, new_realized_and_paper, method = "all")

meanvalue_compute(realized_and_paper, asset_transactions)
}
\arguments{
\item{portfolio_quantity}{Numeric value. The portfolio quantity of the traded asset.}

\item{portfolio_price}{Numeric value. The portfolio price of the traded asset.}

\item{transaction_quantity}{Numeric value. The quantity of the traded asset.}

\item{transaction_price}{Numeric value. The market price of the traded asset.}

\item{transaction_type}{Character string. Either "B" = buy or "S" = sell.}

\item{transaction_datetime}{POSIXct value. The date-time at which the transaction
is going to occur.}

\item{realized_and_paper}{Data frame containing the realized and paper gains and
losses results to be updated.}

\item{new_realized_and_paper}{Data frame containing the realized and paper gains and
losses results related to the new transaction.}

\item{method}{Character string. The method used to compute papers.
Allowed values are "count", "total", "value", "duration" and "all".}

\item{asset_transactions}{Data frame containing the number of transactions for
each asset traded by the investor.}
}
\value{
The described functions have different return behaviours.

\describe{
\item{\code{prz_update}}{returns a numeric value of the new portfolio
price of the traded asset.}
\item{\code{dtt_update}}{returns a datetime value of the new portfolio
datetime of the traded asset.}
\item{\code{results_update}}{returns a \link[tibble:tibble-package]{tibble}
containing the realized and paper gains and losses with the updated
values.}
\item{\code{meanvalue_compute}}{returns a \link[tibble:tibble-package]{tibble}
containing the realized and paper gains and losses with the updated
values.}
}

In particular:

\describe{
\item{\code{RG_"method"}}{contains Realized Gains results.}
\item{\code{RL_"method"}}{contains Realized Losses results.}
\item{\code{PG_"method"}}{contains Paper Gains results.}
\item{\code{PL_"method"}}{contains Paper Losses results.}
}
}
\description{
Unpdate investor's portfolio or realized and paper results
}
\details{

}
\section{Functions}{
\itemize{
\item \code{prz_update}: Update the portfolio price of the traded asset.

\item \code{dtt_update}: Update the portfolio datetime of the traded asset.

\item \code{results_update}: Update the realized and paper gains and losses
results with the results obtained on the last occurred transaction.

\item \code{meanvalue_compute}: Update the realized and paper gains and losses
results averaging the total value by the number of transactions
for each asset.
}}

\references{
H. Shefrin & M. Statman, 1985
}
\seealso{
\code{\link{portfolio_update}}, \code{\link{gains_and_losses}}
}
\author{
L. Mazzucchelli & M. Zanotti
}
