% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{ewise_mean}
\title{Utilities}
\usage{
ewise_mean(x, y, na.rm = FALSE, zero.substitute = FALSE)
}
\arguments{
\item{x, y}{Numeric vectors of the same length.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE}
    indicating whether \code{NA} values should be stripped before the
    computation proceeds.}

\item{zero.substitute}{Logical. If TRUE zero values are substituted with
corresponding non-missing values whether possible.}
}
\value{
Numeric vector.
}
\description{
Utilities functions
}
\section{Functions}{
\itemize{
\item \code{ewise_mean}: Element-wise mean calculation.
}}

\keyword{internal}
